requirements.txt
django-simple-rest==1.4.1
requests==2.25.1

Django~=3.1.7
djangorestframework~=3.12.4
django-crispy-forms~=1.11.2
django-formset-js~=0.5.0
django-ckeditor~=6.0.0
pilkit~=2.0
Pillow~=8.2.0
django-mptt~=0.12.0
django-haystack~=3.0
django-easy-select2~=1.5.7
six~=1.15.0
django-parler~=2.2
modeltranslation~=0.25
django-modeltranslation~=0.16.2